Fix freezeOnCull bug in osgParticle
The _last_frame (used to detect culled particle systems) in osgParticle::ParticleSystem is updated in the draw thread, which can cause issues if the draw hasn't completed by the time the next frame's cull is taking place. In that case the _last_frame will lag another frame behind, causing the particle updater to think the particle system is culled, even if it's visible. This was causing the freezeOnCull option to incorrectly freeze particle systems that are not culled.