osgParticle: override vector interpolations in LinearInterpolator for a small performance boost
This commit and the previous 2 commits resulted in ~300% faster ParticleSystem::update for OpenMW (in Vivec, previously 11% of frame loop, now 4%). Further improvements are possible but we need to get rid of the "interpolator" concept for that, best saved for the new particle system component that will not be based on osgParticle.